// packages/aq-legacy-framework/src/shared/components/layout/CustomBasicAppShell/types.ts
import { ReactNode } from "react";

export interface I_BasicAppShell_LinkItem {
    pageId?: number;
    name?: string;
    label: string;
    status?: "Default" | "Prototype" | "New" | "Menu" | "Change";
    link?: string;
    linkPrototype?: string;
    links?: I_BasicAppShell_LinkItem[];
    note?: string;
    icon?: ReactNode;
    isHidden?: boolean;
    isHideFromSearch?: boolean;
    parentPageId?: number;
}

export interface CustomBasicAppShellProps {
    isDev?: boolean
    children: React.ReactNode;
    menu: I_BasicAppShell_LinkItem[];
    extraTopRight?: React.ReactNode;
    title?: string;
    logoutRedirect?: string;
    /** Tắt gọi API pageService.getAll() (page content) nếu true */
    disablePageContentQuery?: boolean;
    /** Tắt gọi API GetAQModule (logo, favicon, module name) nếu true */
    disableGetAQModuleQuery?: boolean;
}